+++ src/lib.rs
@@ -5,3 +5,4 @@ pub mod perf_events;
 pub mod profiler;
 pub mod unwind_info;
 pub mod usym;
+pub mod collector;
