+++ src/main.rs
@@ -10,7 +10,7 @@ use tracing_subscriber::fmt::format::FmtSpan;
 use tracing_subscriber::FmtSubscriber;
 
 use lightswitch::object::build_id;
-use lightswitch::profiler::Collector;
+use lightswitch::collector::Collector;
 use lightswitch::profiler::Profiler;
 use lightswitch::unwind_info::{compact_printing_callback, UnwindInfoBuilder};
 use primal::is_prime;
